requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
Pillow==10.2.0
Werkzeug==3.0.1
Flask-Limiter==3.5.1
redis==5.0.3 